# Unroll SHA-1 code with a significant speedup