There are two commits in this pull request: first is a plain unroll, second is a small modification to the round function that gives considerable boost on newer CPUs.

Some benchmarkings:


build:
$ ./configure --disable-openssl --enable-static
$ make

build for 32bit:
$ ./configure --disable-openssl --enable-static --extra-cflags=-m32 --extra-ldflags=-m32
$ make


test:
$ dd if=/dev/zero bs=128k count=8k status=progress | taskset 0x01 ./rhash --sha1 -

tests are made with original rhash (orig), unrolled (unroll) and with modified rounds (mod), the numbers are not very accurate and are only provided to compare unrolled and modified code with original.

```
CPU: ryzen 9 5950x, 64bit, gcc 11.2.0
orig:   478 Mb/s
unroll: 990 Mb/s
mod:    1.0 Gb/s

CPU: ryzen 9 5950x, 32bit, gcc 11.2.0
orig:   475 Mb/s
unroll: 875 Mb/s
mod:    920 Mb/s


CPU: amd fx 8320, 64bit, gcc 9.3.0
orig:   200 Mb/s
unroll: 351 Mb/s
mod:    353 Mb/s

CPU: amd fx 8320, 32bit, gcc 9.3.0
orig:   181 Mb/s
unroll: 305 Mb/s
mod:    303 Mb/s


CPU: amd turion II mobile M520, 64bit, gcc 9.3.0
orig:   124 Mb/s
unroll: 230 Mb/s
mod:    228 Mb/s

CPU: amd turion II mobile M520, 32bit, gcc 9.3.0
orig:   133 Mb/s
unroll: 208 Mb/s
mod:    210 Mb/s


CPU: ryzen 5 5600H, 64bit, gcc 11.2.0
orig:   406 Mb/s
unroll: 825 Mb/s
mod:    850 Mb/s

CPU: ryzen 5 5600H, 32bit, gcc 11.2.0
orig:   404 Mb/s
unroll: 746 Mb/s
mod:    770 Mb/s


CPU: intel core i7-3770, 64bit, gcc 8.3.0
orig:   255 Mb/s
unroll: 507 Mb/s
mod:    515 Mb/s


CPU: intel atom D2700, 64bit, gcc 9.3.0
orig:   59.6 Mb/s
unroll: 103 Mb/s
mod:    103 Mb/s

CPU: intel atom D2700, 32bit, gcc 9.3.0
orig:   56.2 Mb/s
unroll: 90.5 Mb/s
mod:    90.7 Mb/s


CPU: cortex-a53, 64bit, clang 13.0.0 
orig:    52 Mb/s
unroll: 108 Mb/s
mod:    106 Mb/s

CPU: cortex-a73, 64bit, clang 13.0.0 
orig:   101 Mb/s
unroll: 180 Mb/s
mod:    180 Mb/s


CPU: intel xeon e5-2680 v3, 64bit, gcc 8.4.0
orig:   214 Mb/s
unroll: 428 Mb/s
mod:    442 Mb/s


CPU: intel core i3-5010U, 64bit, gcc 11.2.0
orig:   137 Mb/s
unroll: 270 Mb/s
mod:    273 Mb/s

CPU: intel core i3-5010U, 32bit, gcc 11.2.0
orig:   131 Mb/s
unroll: 242 Mb/s
mod:    244 Mb/s


CPU: intel celeron M 353, 32bit, gcc 7.5.0
orig:    46 Mb/s
unroll:  73 Mb/s
mod:     72 Mb/s

```