@rhash 2022-04-16
@lvd2, thank you very much for this research and for the significant SHA1 speed up!

The PR is accepted as 0bde67caf64a4e298051c32eb5c635a264c1b005 with small changes in spaces.